requirements.txt
# ==============================================================================
# Tim1-DFK: Requirements for CPT & SFT Pipeline
# ==============================================================================
# Install: pip install -r requirements.txt
# Note: For GPU training, install unsloth separately via:
#   pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"
# ==============================================================================

# --- Core ML ---
torch>=2.1.0
transformers>=4.45.0
datasets>=2.18.0
accelerate>=0.30.0
peft>=0.12.0
trl>=0.9.0
bitsandbytes>=0.43.0

# --- Data Processing ---
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
regex>=2023.0

# --- Monitoring & Experiment Tracking ---
wandb>=0.16.0
tensorboard>=2.14.0

# --- Configuration ---
pyyaml>=6.0
jsonlines>=4.0.0

# --- Utilities ---
tqdm>=4.66.0
rich>=13.0.0
hf_transfer>=0.1.4
